+++ test/unit/fuzz-tools.test.js
@@ -145,6 +145,7 @@ import * as orthoV144 from '../../lib/ortho-v144.js';
 import * as orthoV145 from '../../lib/ortho-v145.js';
 import * as spineV146 from '../../lib/spine-v146.js';
 import * as rheumV147 from '../../lib/rheum-v147.js';
+import * as rheumV148 from '../../lib/rheum-v148.js';
 import * as growthLmsData from '../../lib/growth-lms-data.js';
 import * as emsV149 from '../../lib/ems-v149.js';
 
@@ -236,6 +237,7 @@ const MODULES = {
   'ortho-v145.js': orthoV145,
   'spine-v146.js': spineV146,
   'rheum-v147.js': rheumV147,
+  'rheum-v148.js': rheumV148,
   'growth-lms-data.js': growthLmsData,
   'ems-v149.js': emsV149,
 };